feat: #WB2-1448, add modal linker + refacto
